Add files via upload
Bus Pass and Online Ticket System can be used to book tickets on the go and can be used to
apply/renew the bus pass through smart phone, which helps all generation people. This system
provides connection between government server, where data is managed and android app,
which provides a GUI for the user.